use default java sockets
