+++ rtmp/src/main/java/com/pedro/rtmp/amf/v0/AmfString.kt
@@ -16,8 +16,8 @@
 
 package com.pedro.rtmp.amf.v0
 
+import com.pedro.common.readUntil
 import com.pedro.rtmp.utils.readUInt16
-import com.pedro.rtmp.utils.readUntil
 import com.pedro.rtmp.utils.writeUInt16
 import java.io.IOException
 import java.io.InputStream
